Create engine test fixture and move fixtures into conftest.py
